Fixed a crash in "Implied Dummy Reads"
The implied dummy reads test would crash if ran on a top-loader during the all-test mode. The JSR instruction moved to the wrong address, ($421) so the rest results from the Opcode $47 test results were where it jumped. I have since relocated those test results, and left address $421 as the value $00.